+++ R/venndir-base-polyclip.R
@@ -342,7 +342,7 @@ venndir <- function
  verbose=FALSE,
  debug=0,
  circle_nudge=NULL,
- lwd=1,
+ lwd=0.3,
  rotate_degrees=0,
  ...)
 {