+++ R/venndir-textvenn.R
@@ -72,13 +72,13 @@
 #' # larger number of items
 #' setlist <- make_venn_test(n_items=1000000, sizes=200000, do_signed=TRUE)
 #' # text Venn with directionality
-#' textvenn(setlist, sets=c(1,2,3), "agreement")
+#' textvenn(setlist, sets=c(1,4), "agreement")
 #' 
 #' # basic text Venn with directionality
-#' textvenn(setlist, sets=c(1,2,3), "each")
+#' textvenn(setlist, sets=c(3,4), "each")
 #' 
 #' # simple way to show legend
-#' venndir_legender(textvenn(setlist, overlap_type="each"),
+#' venndir_legender(textvenn(setlist, sets=c(1,2,4), overlap_type="overlap"),
 #'    combine_size=FALSE,
 #'    legend_style="data.frame")
 #' 
@@ -86,7 +86,10 @@
 textvenn <- function
 (setlist,
  sets=seq_along(setlist),
- overlap_type=c("concordance", "overlap", "each", "agreement"),
+ overlap_type=c("concordance",
+    "overlap",
+    "each",
+    "agreement"),
  set_colors=NULL,
  spacing=5,
  padding=1,